+++ CHANGELOG.md
@@ -22,6 +22,8 @@
     - ``LineEditor#bind`` method accept human-readable keyname (such as `F1`, `ctrl+alt+a`, `bs`)
     - ``LineEditor#bindings`` method return human-readable keyname
     - ``insert-keycode`` action no longer accept ``CSI 200 ~`` (bracketed paste start)
+    - recognize kitty keyboard protocol
+- add ``keyboard-protocol`` config to ``LineEditor``. now can enable kitty keyboard protocol
 
 #### Module